Falling block detection and ground light up
I made the blocks detect if they will collide with something in the next movement and stop. I also added a way for the ground tiles to light up if a block is above them so the player can tell where the block will land. There are still a lot of bugs that I won't fix now such as the blocks not detecting the ground in the direction.